refactor(hipdnn): drop frontend pass-by-value execute filter; leave presence to the provider
Remove the execute-time pass-by-value variant-pack filter from the frontend.
It only ever presence-checked pure runtime user-supplied scalars, duplicating
the provider's findDeviceBuffer(uid) guarantee (same INVALID_VALUE, one layer
up) and holding pass-by-value scalars to a stricter standard than every other
tensor — the frontend does no presence-validation for ordinary required
tensors on the execute path, and the backend VariantDescriptor::finalize()
checks only pack self-consistency, not graph-required-UID coverage. A missing
pure-runtime scalar is now caught by the provider exactly like any other
missing tensor.

Also folds in the test-quality items from review:
- Collapse the six near-identical EnsureTensorDescSetsPassByValue* strict
  tests into one TYPED_TEST over the full ValueVariant type set (adds int64/
  bool strict coverage previously only in round-trip).
- Harden the StoringBackend fake: loud ADD_FAILURE on an unhandled unitSize
  type, and clamp the get-path memcpy to the caller's requested capacity.
- Drop the redundant double->float->double casts on the HALF/BFLOAT16 arms of
  resolveScalarOperand (readHostScalar already returns double via the types'
  explicit operator double()).

Removed:
- detail::validatePassByValueVariantPack and its three Graph.hpp execute-path
  callsites (primary execute, execute_plan_at_index, override execute).
- TestPassByValueExecuteFilter.cpp (unit tests of the removed helper) + CMake
  entry.
- The filter-gap TestGraph tests and the no-arg injectValidCompiledPlan helper
  added for them (the 3-arg overload used elsewhere stays).

RFC 0016 §4.9 rewritten to "no frontend filter; the provider is the uniform
backstop" with a rationale subsection; every "override rejected" cross-ref
corrected to "inert/ignored" (the frontend never rejected baked-value entries
— the pack was always forwarded whole and the provider ignores baked-value
slots), and §7.2/§8 Step 5/§9/glossary reconciled to the provider-backstop
model.

Verified: providers + hip-kernel unit targets build clean; frontend/backend/
plugin_sdk/flatbuffers suites all green (1891/355/189/68/133), hip-kernel unit
363 (+1 pre-existing skip). pre-commit clean on all changed files.

+++ projects/hipdnn/plugin_sdk/include/hipdnn_plugin_sdk/RuntimePassByValue.hpp
@@ -112,11 +112,9 @@ inline double resolveScalarOperand(const ScalarOperand& op,
     case DataType::FLOAT:
         return detail::readHostScalar<float>(ptr);
     case DataType::HALF:
-        return static_cast<double>(
-            static_cast<float>(detail::readHostScalar<hipdnn_data_sdk::types::half>(ptr)));
+        return detail::readHostScalar<hipdnn_data_sdk::types::half>(ptr);
     case DataType::BFLOAT16:
-        return static_cast<double>(
-            static_cast<float>(detail::readHostScalar<hipdnn_data_sdk::types::bfloat16>(ptr)));
+        return detail::readHostScalar<hipdnn_data_sdk::types::bfloat16>(ptr);
     case DataType::INT32:
         return detail::readHostScalar<int32_t>(ptr);
     case DataType::INT64:
